requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-ninja==0.19.1
django-on-heroku==1.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
whitenoise==6.2.0
